Move to separate files and remove maps
